Changed uint16_t to 32 and other small changes

+++ src/node.cpp
@@ -6,11 +6,11 @@
 
 #include "node.hpp"
 
-std::unique_ptr<Node> createNode(uint16_t id, char letter) {
+std::unique_ptr<Node> createNode(uint32_t id, char letter) {
     return std::unique_ptr<Node>(new Node(id, letter));
 }
 
-Node::Node(uint16_t id, char letter) :
+Node::Node(uint32_t id, char letter) :
         id_(id), letter_(letter), in_edges_(), out_edges_(), aligned_nodes_ids_() {
 }
