+++ src/node.hpp
@@ -10,7 +10,7 @@
 #include <memory>
 
 class Node;
-std::unique_ptr<Node> createNode(uint16_t id, char letter);
+std::unique_ptr<Node> createNode(uint32_t id, char letter);
 
 class Edge;
 using EdgeSharedPtr = std::shared_ptr<Edge>;
@@ -20,7 +20,7 @@ public:
 
     ~Node();
 
-    uint16_t id() const {
+    uint32_t id() const {
         return id_;
     }
 
@@ -44,27 +44,27 @@ public:
         out_edges_.emplace_back(edge);
     }
 
-    const std::vector<uint16_t>& aligned_nodes_ids() const {
+    const std::vector<uint32_t>& aligned_nodes_ids() const {
         return aligned_nodes_ids_;
     }
 
-    void add_aligned_node_id(uint16_t id) {
+    void add_aligned_node_id(uint32_t id) {
         aligned_nodes_ids_.emplace_back(id);
     }
 
-    friend std::unique_ptr<Node> createNode(uint16_t id, char letter);
+    friend std::unique_ptr<Node> createNode(uint32_t id, char letter);
 
 private:
 
-    Node(uint16_t id, char letter);
+    Node(uint32_t id, char letter);
     Node(const Node&) = delete;
     const Node& operator=(const Node&) = delete;
 
-    uint16_t id_;
+    uint32_t id_;
     char letter_;
 
     std::vector<EdgeSharedPtr> in_edges_;
     std::vector<EdgeSharedPtr> out_edges_;
 
-    std::vector<uint16_t> aligned_nodes_ids_;
+    std::vector<uint32_t> aligned_nodes_ids_;
 };
